requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
pydantic==2.5.0
cors==1.0.1
fastapi-cors==0.0.6